requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
abseil-cpp=20211102.0=hd4dd3e8_0
absl-py=1.0.0=pypi_0
aiohttp=3.8.3=py39h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
asttokens=2.0.5=pypi_0
async-timeout=4.0.2=py39h06a4308_0
attrs=22.1.0=py39h06a4308_0
backcall=0.2.0=pypi_0
blas=1.0=mkl
blinker=1.4=py39h06a4308_0
bottleneck=1.3.4=py39hce1f21e_0
brotli=1.0.9=he6710b0_2
brotlipy=0.7.0=py39h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.0=h5eee18b_0
ca-certificates=2023.08.22=h06a4308_0
cachetools=5.0.0=pypi_0
cattrs=1.5.0=pypi_0
certifi=2023.7.22=py39h06a4308_0
cffi=1.15.0=py39hd667e15_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
cloudpickle=2.0.0=pypi_0
cmake=3.26.1=pypi_0
colorama=0.4.6=py39h06a4308_0
coverage=6.3.2=py39h7f8727e_0
cryptography=37.0.1=py39h9ce1e76_0
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cudatoolkit=11.3.1=h2bc3f7f_2
cycler=0.11.0=pyhd3eb1b0_0
cython=0.29.28=pypi_0
d3rlpy=2.0.4=pyhd8ed1ab_0
d4rl=1.1=pypi_0
dataclasses=0.8=pyh6d0b6a4_7
dataclasses-json=0.5.7=py39h06a4308_0
dbus=1.13.18=hb2f20db_0
decorator=5.1.1=pypi_0
dm-control=1.0.15=pypi_0
dm-env=1.6=pypi_0
dm-tree=0.1.8=pypi_0
etils=1.5.2=pypi_0
executing=0.8.3=pypi_0
expat=2.4.4=h295c915_0
farama-notifications=0.0.4=pypi_0
fasteners=0.17.3=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.10.7=pypi_0
flopth=0.1.3=pypi_0
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=h70c0345_0
frozenlist=1.3.3=py39h5eee18b_0
fsspec=2023.10.0=pypi_0
functorch=2.0.0=pypi_0
giflib=5.2.1=h7b6447c_0
glfw=2.5.3=pypi_0
glib=2.69.1=h4ff587b_1
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
google-api-core=2.11.1=pypi_0
google-auth=2.22.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-cloud-core=2.3.3=pypi_0
google-cloud-storage=2.5.0=pypi_0
google-crc32c=1.5.0=pypi_0
google-resumable-media=2.5.0=pypi_0
googleapis-common-protos=1.60.0=pypi_0
grpc-cpp=1.48.2=h5bf31a4_0
grpcio=1.44.0=pypi_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
gym=0.23.1=pypi_0
gym-notices=0.0.6=pypi_0
gym-unity=0.28.0=pypi_0
gymnasium=0.29.1=pypi_0
gymnasium-robotics=1.2.2=pypi_0
h5py=3.9.0=py39he06866b_0
hdf5=1.12.1=h69dfa17_1
icu=58.2=he6710b0_3
idna=3.3=pyhd3eb1b0_0
imageio=2.18.0=pypi_0
importlib-metadata=4.11.3=pypi_0
importlib-resources=6.1.0=pypi_0
importlib_metadata=6.0.0=hd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
ipdb=0.13.9=pypi_0
ipython=8.2.0=pypi_0
isodate=0.6.1=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.0.3=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h7f8727e_0
kiwisolver=1.3.2=py39h295c915_0
krb5=1.20.1=h568e23c_1
labmaze=1.0.6=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_0
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.7.0.149=0
libcurand=10.3.3.53=0
libcurl=8.2.1=h91b91d3_0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=7.5.0=ha8ba4b0_17
libgfortran4=7.5.0=ha8ba4b0_17
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
libnghttp2=1.52.0=ha637b67_1
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libpng=1.6.37=hbc83047_0
libprotobuf=3.20.3=he621ea3_0
libssh2=1.10.0=h37d81fd_2
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libuv=1.40.0=h7b6447c_0
libwebp=1.2.2=h55f646e_0
libwebp-base=1.2.2=h7f8727e_0
libxcb=1.15=h7f8727e_0
libxml2=2.9.12=h74e7548_1
lit=16.0.0=pypi_0
llvmlite=0.38.1=pypi_0
lxml=4.9.3=pypi_0
lz4-c=1.9.3=h295c915_1
markdown=3.3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.0.1=py39h27cfd23_0
marshmallow=3.19.0=py39h06a4308_0
marshmallow-enum=1.5.1=py39h06a4308_0
matplotlib=3.5.1=py39h06a4308_0
matplotlib-base=3.5.1=py39ha18d171_1
matplotlib-inline=0.1.3=pypi_0
mdurl=0.1.2=pypi_0
minari=0.4.1=pypi_0
mjrl=1.0.0=pypi_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
mlagents=0.28.0=pypi_0
mlagents-envs=0.28.0=pypi_0
mpmath=1.3.0=pypi_0
mujoco=3.0.0=pypi_0
mujoco-py=2.1.2.14=pypi_0
multidict=6.0.2=py39h5eee18b_0
munkres=1.1.4=py_0
mypy_extensions=1.0.0=py39h06a4308_0
ncurses=6.3=h7f8727e_2
nettle=3.7.3=hbbd107a_1
networkx=2.6.3=pyhd3eb1b0_0
numba=0.55.2=pypi_0
numexpr=2.8.1=py39h6abb31d_0
numpy=1.21.5=py39he7a7128_2
numpy-base=1.21.5=py39hf524024_2
nvidia-cublas-cu11=11.10.3.66=pypi_0
nvidia-cuda-cupti-cu11=11.7.101=pypi_0
nvidia-cuda-nvrtc-cu11=11.7.99=pypi_0
nvidia-cuda-runtime-cu11=11.7.99=pypi_0
nvidia-cudnn-cu11=8.5.0.96=pypi_0
nvidia-cufft-cu11=10.9.0.58=pypi_0
nvidia-curand-cu11=10.2.10.91=pypi_0
nvidia-cusolver-cu11=11.4.0.1=pypi_0
nvidia-cusparse-cu11=11.7.4.91=pypi_0
nvidia-nccl-cu11=2.14.3=pypi_0
nvidia-nvtx-cu11=11.7.91=pypi_0
oauthlib=3.2.0=pypi_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1w=h7f8727e_0
packaging=23.1=py39h06a4308_0
pandas=1.4.2=py39h295c915_0
parso=0.8.3=pypi_0
pcre=8.45=h295c915_0
pettingzoo=1.24.0=pypi_0
pexpect=4.8.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.0.1=py39h22f2fdc_0
pip=21.2.4=py39h06a4308_0
portion=2.4.0=pypi_0
prompt-toolkit=3.0.29=pypi_0
protobuf=4.24.2=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=pypi_0
pybullet=3.2.5=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyg=2.0.4=py39_torch_1.11.0_cu113
pygame=2.1.2=pypi_0
pygments=2.16.1=pypi_0
pyjwt=2.4.0=py39h06a4308_0
pynndescent=0.5.7=pypi_0
pyopengl=3.1.7=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py39h2531618_6
pysocks=1.7.1=py39h06a4308_0
python=3.9.7=h12debd9_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-louvain=0.15=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytorch=1.11.0=py3.9_cuda11.3_cudnn8.2.0_0
pytorch-cluster=1.6.0=py39_torch_1.11.0_cu113
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pytorch-scatter=2.0.9=py39_torch_1.11.0_cu113
pytorch-sparse=0.6.13=py39_torch_1.11.0_cu113
pytorch-spline-conv=1.2.1=py39_torch_1.11.0_cu113
pytz=2021.3=pyhd3eb1b0_0
pyyaml=6.0=py39h7f8727e_1
qt=5.9.7=h5867ecd_1
rdflib=6.1.1=pypi_0
re2=2022.04.01=h295c915_0
readline=8.1.2=h7f8727e_1
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.1=pypi_0
rich=13.5.2=pypi_0
rsa=4.8=pypi_0
scikit-learn=1.0.2=py39h51133e4_1
scipy=1.7.3=py39hc147768_0
setuptools=61.2.0=py39h06a4308_0
shellingham=1.5.3=pypi_0
sip=4.19.13=py39h295c915_0
six=1.16.0=pyhd3eb1b0_1
sortedcontainers=2.4.0=pypi_0
sqlite=3.38.3=hc218d9a_0
stable-baselines3=1.6.0=pypi_0
stack-data=0.2.0=pypi_0
structlog=23.1.0=pyhd8ed1ab_0
sympy=1.11.1=pypi_0
tabulate=0.9.0=pypi_0
tensorboard=2.8.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=py39h06a4308_0
termcolor=2.3.0=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.11=h1ccaba5_1
toml=0.10.2=pypi_0
torch=2.0.0=pypi_0
torchaudio=0.11.0=py39_cu113
torchdiffeq=0.2.3=pypi_0
torchvision=0.12.0=py39_cu113
tornado=6.1=py39h27cfd23_0
tqdm=4.66.1=pypi_0
traitlets=5.1.1=pypi_0
triton=2.0.0=pypi_0
typer=0.9.0=pypi_0
typing-extensions=4.4.0=pypi_0
typing_extensions=4.7.1=py39h06a4308_0
typing_inspect=0.9.0=py39h06a4308_0
tzdata=2022a=hda174b7_0
umap-learn=0.5.3=pypi_0
urllib3=1.26.9=py39h06a4308_0
wcwidth=0.2.5=pypi_0
werkzeug=2.0.3=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.5=h7f8727e_1
yacs=0.1.6=pyhd3eb1b0_1
yaml=0.2.5=h7b6447c_0
yarl=1.8.1=py39h5eee18b_0
zipp=3.7.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.4.9=haebb681_0
